Validate alignment of SizeOfOptionalHeader sooner
It is technically redundant, but it means that optional_header_size
really has been validated, and is much less brittle if the code changes
in the future.  For instance, one can obtain the pointer to the section
headers without needing to first check that the NumberOfRvaAndSizes
field is correct.  Also, the earlier validation gives a much better
error message.