Revert "Allow extra data between sections and signature"
It turns out that this data _should_ be part of the signature, but in
many implementations (notably EDK2) it is _not_.  This is a bug in these
implementations, but it is simplest to just disallow this nasty case.

In the unsigned case, continue to allow data after the last section.

This reverts commit 191a401e073d7a372372f7f99e50e099562794bc.