Add flags to control strictness and whether relocations must be preserved
This allows disabling checks that don't to pass for a file to be valid,
but which are needed for things like secure boot or loading the image at
an address other than what it was linked for.  It also allows ensuring
that the .reloc section is not marked as discardable, which causes Xen
to crash with some bootloaders.